Widgets folder - A folder for widgets has been created. - One class, that extends an Inherited Widget, was added to 'widgets' folder.
