+++ lib/screens/profile_screen.dart
@@ -1,7 +1,7 @@
-import 'package:delivery/screens/main_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:sliding_sheet/sliding_sheet.dart';
+import 'package:delivery/widgets/constant_inherited_widget.dart';
 
 class Profile extends StatefulWidget {
   const Profile({Key? key}) : super(key: key);
